Don't use the etcd embedded client
It seems that in some circumstances, the etcd embedded client can
cause nil panics. For instance, when the embedded etcd is shut
down, the client cannot be safely used anymore.

Since this is an unacceptable state of affairs, we will revert to
using the gRPC client, which seems to be much more reliable.

Signed-off-by: Eric Chlebek <eric@sensu.io>